SLUG/SHARE/FB_COMMENT_LIKE/USER_PERMISSION!

+++ landing/views.py
@@ -1,6 +1,7 @@
+from urllib.parse import quote
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 
 from .forms import PostForm
@@ -8,9 +9,12 @@ from .models import Post
 
 
 def post_create(request):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         # message success
         messages.success(request, "Successfully Created")
@@ -21,11 +25,13 @@ def post_create(request):
     return render(request, "post_form.html", context)
 
 
-def post_detail(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_detail(request, slug=None):
+    instance = get_object_or_404(Post, slug=slug)
+    share_string = quote(instance.content)
     context = {
         "title": instance.title,
         "instance": instance,
+        "share_string": share_string,
     }
     return render(request, "post_detail.html", context)
 
@@ -52,8 +58,10 @@ def post_list(request):
     return render(request, "post_list.html", context)
 
 
-def post_update(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_update(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
         instance = form.save(commit=False)
@@ -69,8 +77,10 @@ def post_update(request, id=None):
     return render(request, "post_form.html", context)
 
 
-def post_delete(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_delete(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Successfully deleted")
     return redirect("posts:list")
